`storage`: move `storage_e2e_fixture` to its own header file
(cherry picked from commit c0d76d1afbccecc6762efc365d185b70feb3fdb6)
